+++ mcp_tools/binary_analysis/binary_analysis_and_reverse_engineering.py
@@ -3,85 +3,6 @@
 from typing import Dict, Any
 
 def register_binary_analysis_and_reverse_engineering_tools(mcp, hexstrike_client, logger):
-    @mcp.tool()
-    def gdb_analyze(binary: str, commands: str = "", script_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute GDB for binary analysis and debugging with enhanced logging.
-
-        Args:
-            binary: Path to the binary file
-            commands: GDB commands to execute
-            script_file: Path to GDB script file
-            additional_args: Additional GDB arguments
-
-        Returns:
-            Binary analysis results
-        """
-        data = {
-            "binary": binary,
-            "commands": commands,
-            "script_file": script_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting GDB analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/gdb", data)
-        if result.get("success"):
-            logger.info(f"✅ GDB analysis completed for {binary}")
-        else:
-            logger.error(f"❌ GDB analysis failed for {binary}")
-        return result
-
-    @mcp.tool()
-    def radare2_analyze(binary: str, commands: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Radare2 for binary analysis and reverse engineering with enhanced logging.
-
-        Args:
-            binary: Path to the binary file
-            commands: Radare2 commands to execute
-            additional_args: Additional Radare2 arguments
-
-        Returns:
-            Binary analysis results
-        """
-        data = {
-            "binary": binary,
-            "commands": commands,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Radare2 analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/radare2", data)
-        if result.get("success"):
-            logger.info(f"✅ Radare2 analysis completed for {binary}")
-        else:
-            logger.error(f"❌ Radare2 analysis failed for {binary}")
-        return result
-
-    @mcp.tool()
-    def binwalk_analyze(file_path: str, extract: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Binwalk for firmware and file analysis with enhanced logging.
-
-        Args:
-            file_path: Path to the file to analyze
-            extract: Whether to extract discovered files
-            additional_args: Additional Binwalk arguments
-
-        Returns:
-            Firmware analysis results
-        """
-        data = {
-            "file_path": file_path,
-            "extract": extract,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Binwalk analysis: {file_path}")
-        result = hexstrike_client.safe_post("api/tools/binwalk", data)
-        if result.get("success"):
-            logger.info(f"✅ Binwalk analysis completed for {file_path}")
-        else:
-            logger.error(f"❌ Binwalk analysis failed for {file_path}")
-        return result
 
     @mcp.tool()
     def ropgadget_search(binary: str, gadget_type: str = "", additional_args: str = "") -> Dict[str, Any]:
